SONiC landing page "https://azure.github.io/SONiC/" build error due to this % usage in md
SONiC webpage at https://azure.github.io/SONiC/ is not loading properly with the modified index.html page. It gives "build error". While trying to find the root cause of this error, it gave the following error in pmon design doc.
<<Liquid Exception: Liquid error (line 311): Unknown operator skip_ledd in doc/pmon/pmon-enhancement-design.md>>
We suspect that repository gave build error due to this line 311 error.
In general, in markdown, if we want to use "%" kind of special characters, we need to give a "\" before such characters. 
When this "\" added in this line 311, the "Liquid Exception" error is not seen. 
Webpage build error can be verified only after this change is checked in.
We will merge this change and see if webpage build error is going away. If this is an incorrect change, we will revert this change later.