usage of % in the file needs a backslash before %
SONiC webpage at https://azure.github.io/SONiC/ is not loading properly with the modified index.html page. It gives "build error" from GitHub that says “The tag `if` on line 185 in `doc/sonic-build-system/build_system_improvements.md`”. 
While trying to find the root cause of this error, we found the file build_system_improvements.md is giving error due to usage of “%” in markdown file.
We suspect that repository gave build error due to the lines 20, 22 & 24.
In general, in markdown, if we want to use "%" kind of special characters, we need to give a "\" before such characters. 
When this "\" added in those lines, 
Webpage build error can be verified only after this change is checked in.
We will merge this change and see if webpage build error is going away. If this is an incorrect change, we will revert this change later.