for loop inside ACL_RULE gives github page build failure - added backslash as a temporary solution
While trying to check index.html file in the repo, github gives page build failure message that stated "There is a syntax error in the 'for' loop on line 386 of `doc/vrf/vrf-ansible-test-plan.md`. Valid syntax is 'for <item> in <collection>', and the <item> variable may not contain hyphens or quotation marks. "
This is due to the usage of "percentage " and "for" which are being parsed as part of github build process where it checks for syntax.
As a workaround, a backslash is added in front of all "%" symbols and a NOTE has been added to explain the same.  Module owners are requested to correct the right markdown format for this.