Correct a duplicated unit test case to specify the correct test description
The "Unit Test" section of the document is intended to specify two invalid "date" cases. One is an invalid date. The other (applicable only to the CLI client interface) is typing the "date" sub-command/keyword without following it with text specifying a Date/time value ("no" date string after entering the "date" sub-command mode). The document previously contained two instances of the first of these two cases and was missing the second one.

I have corrected this mistake by editing the second copy of the "invalid date" case to specify that the testing should cover the "no date" case.