SONiC configuration setup enhancements to provide installer migration hooks capability
The following new features are being added to the config-setup service:
1. Provide a second set of configuration pre hooks as part of the installer
image. This allows the user to fix issues caused by the migration-pre
scripts already installed on the switch.

2. Both the config-migration-pre and config-migration-post are migrated to
the newly installed image. This allows end user installed scripts to be
carried over to the newly installed image automatically to make it easier
for subsequent configuration migrations.

Signed-off-by: Rajendra Dendukuri <rajendra.dendukuri@broadcom.com>