webpage build error - Add extra whitespace
SONiC landing page "https://azure.github.io/SONiC/" build error due to the usage of % in markdown md files.

SONiC webpage at https://azure.github.io/SONiC/ is not loading properly with the modified index.html page. It gives "build error". While trying to find the root cause of this error, it gave the following error in this design doc.
<<Liquid Exception: Liquid syntax error (line 139): Unknown tag 'set' in doc/database/multi_namespace_db_instances.md>>
This "page build error" is due to this line 139.
In general, in markdown, if we want to use "%" kind of special characters, we need to give a whitespace before such characters. 
When the whitespace got added in this line 139, the "Liquid Exception" error is not seen. 
Webpage build error disappears after this change.