HLD for the SONiC Management Framework "show tech-support" command (#756)
This initial version of the SONiC Management Framework "show tech-support" command describes the SONiC Management Framework interface for initiating the collection of technical support information. The specification describes the associated Yang model and flow of control for relaying the command to the SONiC host and receiving a completion status and the location of the resulting techsupport "bundle".