HLD update: Adding support for FPGA devices in PDDF (#1232)
HW vendors use FPGA devices in their platform design.

In one of the use cases, the FPGA is connected to CPU via I2C bus. Such FPGAs are similar to any other I2C device with I2C address and adapter where they are connected. We are calling it FPGAI2C.

The other use case is where the FPGA is connected to CPU via PCIe. The FPGA also acts as the I2C controller and other devices are controlled by it. We are calling this design as FPGAPCIe.

This update in the PDDF HLD deals with the support for these two kind of FPGAs.

Co-authored-by: Ying Xie <yxieca@users.noreply.github.com>