[system-health] [update] Add fan direction check to system health (#1312)
- What I did?
Add a fan direction check to system health.

- Why I did this?
Correct airflow is very important for a switch system to prevent it from overheating. Incorrect airflow would cause abnormal system temperature raising and it is difficult to debug it without open the switch box.

As fan direction is already queried and stored to DB by thermalctld, system-health service should monitor fan direction and raise alarm if fan direction is not correct.